Initial Commit
Grabbed some example code for setting up a simple snake game with tkinter and heavily modified it. Basically completely re-writing it. Just needed an example usage of tkinter that we could test. Set up a gamewindow class and snake class to set up a window to play the game in and to create snakes, respectively